Drop the `ssize_t` cast at the sites whose index already is `ssize_t`
The `offsets[i - 1]` -> `offsets[static_cast<ssize_t>(i) - 1]` rewrite also
landed where the index is already signed and the cast is a no-op, which
`clang-tidy` reports as `readability-redundant-casting`.  The two such casts in
`ColumnArray.h` failed 1658 objects on their own, almost every translation unit
including that header.

Remove the cast at those 26 sites; the places where the index is `size_t`, which
are the point of the rewrite, stay as they are.  Four sites in `arrayFold.cpp`
are included that the failing build had not reported yet.

Reported by Build (arm_tidy) on https://github.com/ClickHouse/ClickHouse/pull/120906.

+++ src/Columns/ColumnArray.h
@@ -236,8 +236,8 @@ private:
     WrappedPtr data;
     WrappedPtr offsets;
 
-    size_t ALWAYS_INLINE offsetAt(ssize_t i) const { return getOffsets()[static_cast<ssize_t>(i) - 1]; }
-    size_t ALWAYS_INLINE sizeAt(ssize_t i) const { return getOffsets()[i] - getOffsets()[static_cast<ssize_t>(i) - 1]; }
+    size_t ALWAYS_INLINE offsetAt(ssize_t i) const { return getOffsets()[i - 1]; }
+    size_t ALWAYS_INLINE sizeAt(ssize_t i) const { return getOffsets()[i] - getOffsets()[i - 1]; }
 
     /// Multiply values if the nested column is ColumnVector<T>.
     template <typename T>
